Create a baseclass for `device` and `uinput`, to hold the shared functions that read and write events
